Extracted some html responses logic to abstract class
